build system: add --enable-libdbi=optional functionality
If set, builds a dummy module which just emits a "module not supported
on this platform" error message when loaded.

Primary use case for this system is Debian-ish builds on SUSE OBS,
where we prefer to have a single package definition for all versions
(else things get much more complicated).