remove rsyslog.service[.in] sample file
This file is a systemd unit file. Over the past months, we
received numerous complaints from the RH'ish community because
of the "wrong" content of this file. Thus, we changed it to
silence these complaints. Now, very similar complains come
from the Debian'ish community
(https://github.com/rsyslog/rsyslog/pull/4317#discussion_r445907950).

The root cause of these problems is that this is not really
a ryslog-related file. It is a systemd unit file, and obviously
very highly depending on distro policies. It was an excellent
idea to add this file back in the early days of systemd when
nobody knew about unit files. Nowadays, however, a single
"proposed" unit file causes more trouble than it solves.

As such, we will remove the file in rsyslog's project root.
Instead, we will provide distro-specific sample files in the
./platform subdirectory.

This way each distro can maintain its (considerably different)
rsyslog.service without the rsyslog interfering with it.

closes https://github.com/rsyslog/rsyslog/issues/4333