fmunflatten: rscript function to unflatten keys in a JSON object
This commit adds a new rainerscript function to unflatten keys
in a JSON tree. It provides a way to expand dot separated fields.

<result> = unflatten(<source-tree>, <key-separator-character>);

It allows for instance to produce this:
    { "source": { "ip": "1.2.3.4", "port": 443 } }

from this source data:
    { "source.ip": "1.2.3.4", "source.port": 443 }

Here is a sample use case:

set $!source.ip = "1.2.3.4";
set $!source.bytes = 3258;
set $!source.geo.country_iso_code = "FR";
set $!destination.ip = "4.3.2.1";

set $.unflatten = unflatten($!, ".");

if (script_error() == 0) then {
    unset $!;
    set $! = $.unflatten;
    unset $.unflatten;
}

{
  "source": {
    "ip": "1.2.3.4",
    "bytes": 3258,
    "geo": {
      "country_iso_code": "FR"
    }
  },
  "destination": {
    "ip": "4.3.2.1"
  }
}

This patch should address (as a post-processing workaround) the needs
expressed in rsyslog/liblognorm#105: permit specifying a multi-level
name. I believe this is somehow what @rgerhards suggested in a comment
on the issue.

The patch was originally written to address a lack of functionality in
Elasticsearch ingest pipeline processors. Some processors would not
accept dotted fields as input (or output, can't remember exactly, at
least in 7.8.1). The dot expander processor can be used to unflatten
fields but this is not generic (and thus, difficult to maintain) as
you have to apply the processor on a known list of fields.