imfile: do_inotify() use only one blocking call
do_inotify block only once for events with timeouts. There is possible
scenario where no data is available and read() enters blocking while poll()
is released.

The change introduced here changes the racy reading to be same for the
triggered file as for the rest of the files:

current:
- Inotify is triggered
- Triggering file is read for new data
- Rest of the files are read for timer expiration
- more inotifys are processed these may be at the end of the file already as well then

new:
- Inotify is triggered
- All files are read for timer expiration, which includes the triggering file
- Triggering files is read for data, this may be at the end of the file already then
- more inotifys are processed these may be at the end of the file already as well then

Therefore the change introduced harmonises the way triggering and the future triggers
are handled, by not making the triggering one as an exceptional one, so it is easier
to test the changes as well.