omfwd bugfix: segfault or error if port not given
If omfwd is configured via RainerScript config format and the "port"
parameter is not given, a segfault will most likely happen on
connection establishment for TCP connections. For UDP, this is
usually not the case.

Alternatively, in any case, errors may happen.

Note that the segfault will usually happen right on restart so this
was easy to detect.

We did not receive reports from practice. Instead, we found the bug
while conducting other work.