CI: disable Travis CI for the time being
The Travis definitions and environments are heavily outdated. Also,
Travis is doing changes to their systems.

This is a good time to disable Travis builds, as everything is
covered by other CI runs as well. We need to reconsider if we
update Travis or keep working within the bounds of the rest of
the CI system. In other words: is Travis still worth it?