imudp: add socket type (IPv4 vs. 6) to input name
Most importantly, the input name is used for stats counter names as
well. Previously, the same name was used for IPv4 and IPv6, so we had
two counters with an equal name. That left users puzzled.

Unfortunately, this change can potentially require changes to existing
analysis scripts, as the name is now slightly different.

closes https://github.com/rsyslog/rsyslog/issues/4364