CI: remove fedora 33 based testing
We have some hang conditions solely with Fedora 33, which we do not
see with other Fedora versions. As we already have f 34 inside CI, we
decide to not necessarily analyze this further but rather remove this
environment from CI.

If time permits, we still try to find the root cause, but so far we
are very sure it's not directly in rsyslog.