ompgsql: PGsslInUse not supported on old distros
Check for existence of PGsslInUse and fallback to PGgetssl if not present. Use of
PGgetssl in this way is not supported for libpq 9.5+, but we still need to support
distros that package the older libpq.

This may be a bit overkill just for a debug message, but being able to easily
check connection TLS state client-side is helpful.