omelasticsearch: Fix indexSuccess impstats counter in bulkmode
When bulkmode is enabled, and a batch was processed without any
failures (errors is false), the code that increments the indexSuccess
impstats counter was never reached. As fix, we obtain the numitems
first and add them to the indexSuccess impstats counter if
errors is false.

This fix may solve some of the random elasticsearch testbench failures.

closes: https://github.com/rsyslog/rsyslog/issues/4794