atomic_operations.m4: also test these opts on time_t.
On NetBSD, time_t has for a long time now been __int64_t.
On 32-bit CPUs, the compiler is not obliged to define
__sync_bool_compare_and_swap_8, so instead this ends up
as an undefined symbol when linking rsyslog.  This makes
the code fall back to the pthread / locking method on these
systems, but at least lets the program build.