bugfix: prevent pot. segfault when switchung to queue emergency mode
When switching to Disk queue emergency mode, we destructed the in-memory
queue object.  Practice has shown that this MAY cause races during
destruction which themselfs can lead to segfault. For that reason, we
now keep the disk queueu object. This will keep some ressources,
including disk space, allocated. But  we prefer that over a segfault.
After all, it only happens after a serious queue error when we are
already at the edge of hard problems.

see also: https://github.com/rsyslog/rsyslog/issues/4963