Fix quoting for omprog, improg, mmexternal
This changes the current behaviour from honouring the double quotes in
any part of the argument but leaving them in place and passing to the
executed binary to requiring quotes exactly at the beginning and at the
end of a multi-word argument, and not including them in the actual call.

Testcases added to ensure the expected results.

Fixes #4249.