tcpflood: fix today's programming error
In commit 5050249a1e tcpflood was "fixed" to work properly when senden
data to a TCP stream socket. Unfortunately, there was an undetected
copy&past problem, which this commit corrects.

The issue did unfortunatley not surfcase with the previous commit as
the situation where this can happen occurs infrequently and randomly
(has to do with timing on the system in question). So the original PR
went fine, but later PRs brought the issue up.

This is not a real regression, as the original fix was for the same
symptom, but with different cause. Now everything should indeed be
fixed.

NOTE: same day correction on a testbench tool