tcpflood: add -A option to NOT abort when sending fails
Note: tcpflood is a testbench tool, not meant for production use

We add the -A option. If set, it does NOT abort tcpflood if
sending messages fails.

This is required for some tests which test connection closes.
If tcpflood terminates quickly before rsyslog could fully
initializes all sessions, the expected error message count is
not necessarily reached, which results in flakes. When tcpflood
contines to run, rsyslog has sufficient time (at least a race
is far less likely).