TLS subsystem: add remote hostname to error reporting
This provides richer and easier to process logs for error and warning
cases. One goal is to enable automatic operations without the need
to consolidate multiple message to a single information.

This improves one situation in gtls driver and provides a more
generic approach in ossl driver for OpenSSL error reporting.

There is probably still room for improvement, however this patch
is at least a good starting point for further work. Please
provide feedback if you need more!

closes https://github.com/rsyslog/rsyslog/issues/5244