ommysql bugfix: potential segfault on database error
Due to an invalid code path, ommysql may cause a segfault if database
transactions fail into a specific way. The main trigger is a totally
irrecoverrable database error which can lead to premature connection
close, which is not checked for in all recover code.

This was detected in a setting where a stored procedure is called that
rolls back a transaction in itself.

This patch fixes the issue.

closes https://github.com/rsyslog/rsyslog/issues/5288