omprog: Recommend libexec directory for binaries
The documentation does not explicitly mention, where should
users place their binaries that will be executed by the omprog
module. Whilst rsyslog does not really care where these files
are located as far as permission and ownership are fine, there
are many security hardening utilities in the wild that consider
this as an important aspect.

From Filesystem Hierarchy Standard point of view, definition of
/usr/libexec: Binaries run by other programs that are not
intended to be executed directly by users or shell
scripts (optional).