NET-OSSL driver: Created interfaces for exported functions
Shared net_ossl_* functions are not available if loaded dynamically
by module in imdtls or omdtls. The incorrect workarround was that I linked
lmnsd_ossl.la into omdtls and imdtls. However that breaks make install when
DTLS modules are enabled.

To fix the problem, all net_ossl_ functions are mapped to the net_ossl
interface which is dynamically loaded on runtime.

This removes the lmnsd_ossl.la dependency from the omdtls and imdtls makefiles
and fixes make install.

it also removes this build warning:
*** Warning: Linking the shared library omdtls.la against the loadable module
*** lmnsd_ossl.so is not portable!

closes: https://github.com/rsyslog/rsyslog/issues/5340